+++ odesk/routers/hr.py
@@ -175,8 +175,7 @@ class HR(Namespace):
     """team api"""
 
     def post_team_adjustment(self, team_reference, engagement_reference,
-                             comments, amount=None, charge_amount=None,
-                             notes=None):
+                             comments, charge_amount, notes=None):
         """
         Add bonus to an engagement.
 
@@ -188,11 +187,7 @@ class HR(Namespace):
           :comments:              Comments about this adjustment, e.g.
                                  "Bonus for a good job"
 
-          :amount:                (conditionally optional) The amount that
-                                  the provider should receive, e.g. 100
-
-          :charge_amount:         (conditionally optional) The amount that
-                                  will be charged to the employer, e.g. 110
+          :charge_amount:         The amount that will be charged to the employer, e.g. 110
 
           :notes:                 (optional) Notes
 
@@ -206,17 +201,7 @@ class HR(Namespace):
 
         data['engagement__reference'] = engagement_reference
         data['comments'] = comments
-
-        if (amount and charge_amount) or (amount is None and
-                                          charge_amount is None):
-            raise ApiValueError('Either only one of the parameters ``amount``'
-                                ' or ``charge_amount`` should be specified')
-
-        if amount:
-            data['amount'] = amount
-
-        if charge_amount:
-            data['charge_amount'] = charge_amount
+        data['charge_amount'] = charge_amount
 
         if notes:
             data['notes'] = notes
@@ -351,7 +336,7 @@ class HR(Namespace):
 
     def post_job(self, buyer_team_reference, title, job_type, description,
                  visibility, category, subcategory, budget=None, duration=None,
-                 start_date=None, end_date=None, skills=None):
+                 start_date=None, skills=None):
         """
         Post a job.
 
@@ -402,10 +387,6 @@ class HR(Namespace):
                                    included the job will default to
                                    starting immediately.
 
-          :end_date:               (deprecated) This parameter remains
-                                   for compatibility reasons and
-                                   will be removed in future.
-
           :skills:                 (optional) Skills required for the job.
                                    Must be a list or tuple even of one item,
                                    e.g. ``['python']``
@@ -446,7 +427,7 @@ class HR(Namespace):
 
     def update_job(self, job_id, buyer_team_reference, title, description,
                    visibility, category=None, subcategory=None, budget=None,
-                   duration=None, start_date=None, end_date=None, status=None):
+                   duration=None, start_date=None, status=None):
         """
         Update a job.
 
@@ -494,10 +475,6 @@ class HR(Namespace):
                                    e.g. 06-15-2011. If start_date is not
                                    included the job will default to
                                    starting immediately.
-
-          :end_date:               (deprecated) This parameter remains
-                                   for compatibility reasons and
-                                   will be removed in future.
 
           :status:                 (required) The status of the job,
                                    e.g. 'filled'.